+++ CHANGELOG.md
@@ -21,6 +21,15 @@ Changes are grouped as follows
 ### Added
 - Support for the `/simulators/models` and `/simulators/models/revisions` API endpoints.
 - Support for the `/simulators` and `/simulators/integration` API endpoints.
+
+## [7.73.3] - 2025-02-07
+### Added
+- Listable property types for containers in Data Modeling now accept `max_list_size`.
+
+## [7.73.2] - 2025-02-05
+### Fixed
+- An edge case where instance IDs in failing requests would not be reported in the error attributes `unknown`
+  or `failed`.
 ### Changed
 - Set the limit for create, update, and delete endpoints for postgres gateway users to 1.
 
@@ -30,7 +39,7 @@ Changes are grouped as follows
 
 ## [7.73.0] - 2025-01-22
 ### Added
-- Data Workflows: Support for `useTransformationCredentials` for the transformations task. This allows running 
+- Data Workflows: Support for `useTransformationCredentials` for the transformations task. This allows running
   transformation tasks with pre-configured client credentials.
 
 ## [7.72.2] - 2025-01-20